Add FTP support
* add FTP support with jlaffaye/ftp library, introduce scheme-aware filtering and content-type detection for HTTP(S) only, refactor client/crawler code into separate protocol-specific files (client_http.go, client_ftp.go, crawler_http.go, crawler_ftp.go), and update README to clarify HTTP(S)/FTP scheme support
* skip FTP file tasting when no MIME filters configured by detecting non-directory entries via trailing slash and queueing them directly for download, add test verifying zero fileListCalls with two downloadEntries
