+++ go.mod
@@ -7,6 +7,7 @@ toolchain go1.24.12
 require (
 	github.com/PuerkitoBio/goquery v1.11.0
 	github.com/alexflint/go-arg v1.5.1
+	github.com/jlaffaye/ftp v0.2.0
 	github.com/sirupsen/logrus v1.9.3
 	github.com/stretchr/testify v1.9.0
 	github.com/temoto/robotstxt v1.1.2
@@ -16,6 +17,8 @@ require (
 	github.com/alexflint/go-scalar v1.2.0 // indirect
 	github.com/andybalholm/cascadia v1.3.3 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/hashicorp/errwrap v1.0.0 // indirect
+	github.com/hashicorp/go-multierror v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	golang.org/x/net v0.47.0 // indirect
 	golang.org/x/sys v0.38.0 // indirect
